extracted test logic to helper assertions and protocols
